requirements.txt
pyTelegramBotAPI
python-dotenv
Pillow
requests
google-genai
pydub
librosa
groq
